+++ tests/test_plugins.py
@@ -2,15 +2,15 @@ from __future__ import annotations
 
 import re
 import sys
-from collections import namedtuple
-from dataclasses import dataclass
 from typing import Any
 
 import pytest
 
+from tests.mocked_plugins import MockedDistribution
+from tests.mocked_plugins import MockedEntryPoint
+from tests.mocked_plugins import MockedPluginA
 from variantlib.base import PluginType
 from variantlib.base import VariantFeatureConfigType
-from variantlib.base import VariantPropertyType
 from variantlib.errors import PluginError
 from variantlib.errors import PluginMissingError
 from variantlib.loader import PluginLoader
@@ -23,93 +23,6 @@ if sys.version_info >= (3, 10):
     from importlib.metadata import EntryPoint
 else:
     from importlib_metadata import EntryPoint
-
-
-class MockedPluginA(PluginType):
-    namespace = "test_namespace"
-
-    def get_all_configs(self) -> list[VariantFeatureConfigType]:
-        return [
-            VariantFeatureConfig("name1", ["val1a", "val1b", "val1c", "val1d"]),
-            VariantFeatureConfig("name2", ["val2a", "val2b", "val2c"]),
-        ]
-
-    def get_supported_configs(self) -> list[VariantFeatureConfigType]:
-        return [
-            VariantFeatureConfig("name1", ["val1a", "val1b"]),
-            VariantFeatureConfig("name2", ["val2a", "val2b", "val2c"]),
-        ]
-
-    def get_build_setup(
-        self, properties: list[VariantPropertyType]
-    ) -> dict[str, list[str]]:
-        for prop in properties:
-            assert prop.namespace == self.namespace
-            if prop.feature == "name1":
-                return {
-                    "cflags": [f"-march={prop.value}"],
-                    "cxxflags": [f"-march={prop.value}"],
-                    "ldflags": ["-Wl,--test-flag"],
-                }
-        return {}
-
-
-MyVariantFeatureConfig = namedtuple("MyVariantFeatureConfig", ("name", "values"))
-
-
-# NB: this plugin deliberately does not inherit from PluginType
-# to test that we don't rely on that inheritance
-class MockedPluginB:
-    namespace = "second_namespace"
-
-    def get_all_configs(self) -> list[MyVariantFeatureConfig]:
-        return [
-            MyVariantFeatureConfig("name3", ["val3a", "val3b", "val3c"]),
-        ]
-
-    def get_supported_configs(self) -> list[MyVariantFeatureConfig]:
-        return [
-            MyVariantFeatureConfig("name3", ["val3a"]),
-        ]
-
-
-class MyFlag:
-    name: str
-    values: list[str]
-
-    def __init__(self, name: str) -> None:
-        self.name = name
-        self.values = ["on"]
-
-
-class MockedPluginC(PluginType):
-    namespace = "incompatible_namespace"
-
-    def get_all_configs(self) -> list[VariantFeatureConfigType]:
-        return [
-            MyFlag("flag1"),
-            MyFlag("flag2"),
-            MyFlag("flag3"),
-            MyFlag("flag4"),
-        ]
-
-    def get_supported_configs(self) -> list[VariantFeatureConfigType]:
-        return []
-
-    def get_build_setup(
-        self, properties: list[VariantPropertyType]
-    ) -> dict[str, list[str]]:
-        flag_opts = []
-
-        for prop in properties:
-            assert prop.namespace == self.namespace
-            assert prop.value == "on"
-            flag_opts.append(f"-m{prop.feature}")
-
-        return {
-            "cflags": flag_opts,
-            "cxxflags": flag_opts,
-        }
 
 
 class ClashingPlugin(PluginType):
@@ -139,49 +52,6 @@ class ExceptionTestingPlugin(PluginType):
     def __call__(self) -> ExceptionTestingPlugin:
         """Fake instantiation"""
         return self
-
-
-@dataclass
-class MockedDistribution:
-    name: str
-    version: str
-
-
-@dataclass
-class MockedEntryPoint:
-    name: str | None
-    value: str
-    plugin: Any
-    group: str | None = None
-    dist: MockedDistribution | None = None
-
-    def load(self) -> Any:
-        return self.plugin
-
-
-@pytest.fixture
-def mocked_plugin_loader(session_mocker):
-    session_mocker.patch("variantlib.loader.entry_points")().select.return_value = [
-        MockedEntryPoint(
-            name="test_namespace",
-            value="tests.test_plugins:MockedPluginA",
-            dist=MockedDistribution(name="test-plugin", version="1.2.3"),
-            plugin=MockedPluginA,
-        ),
-        MockedEntryPoint(
-            name="second_namespace",
-            value="tests.test_plugins:MockedPluginB",
-            dist=MockedDistribution(name="second-plugin", version="4.5.6"),
-            plugin=MockedPluginB,
-        ),
-        MockedEntryPoint(
-            name="incompatible_namespace",
-            value="tests.test_plugins:MockedPluginC",
-            plugin=MockedPluginC,
-        ),
-    ]
-
-    return PluginLoader
 
 
 def test_get_all_configs(mocked_plugin_loader: type[PluginLoader]):
